Refactor: PostViewPage 리팩토링
View 부분을 제외한 로직 부분을 클래스화하여 post_view_utils에 저장함.
